fix: add cast to silence warning

+++ src/c/vensim.c
@@ -184,9 +184,9 @@ double* _VECTOR_SORT_ORDER(double* vector, size_t size, double direction) {
     // TODO signal error
     return NULL;
   }
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     d[i].x = vector[i];
-    d[i].ind = i;
+    d[i].ind = (int)i;
   }
   sort_order = direction > 0.0 ? 1 : -1;
   qsort(d, size, sizeof(dbl_ind), __compare_dbl_ind);
